Extract helper methods to a separate file
